@@ -1,6 +1,6 @@
 {
   "name": "react-calendar-timeline",
-  "version": "0.26.7",
+  "version": "0.27.3-rr",
   "description": "react calendar timeline",
   "main": "lib/index.js",
   "scripts": {
@@ -13,7 +13,7 @@
     "prepublish": "npm run build:lib",
     "start": "webpack-dev-server --hot --host 0.0.0.0 --display-modules",
     "test": "jest",
-    "test:watch": "jest --watch"
+    "test:watch": "jest --watch"  
   },
   "files": [
     ".babelrc",
@@ -114,12 +114,15 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.5.2",
     "@babel/preset-env": "^7.5.2",
     "@babel/preset-react": "^7.0.0",
+    "@testing-library/jest-dom": "^4.2.0",
+    "@types/jest": "24",
     "babel-eslint": "^7.1.1",
     "babel-jest": "^24.8.0",
     "babel-loader": "^8.0.6",
     "babel-plugin-react-remove-properties": "^0.3.0",
     "cross-env": "^5.1.4",
     "css-loader": "~0.26.0",
+    "d3": "^5.12.0",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.14.0",
     "eslint": "^4.16.0",
@@ -133,7 +136,6 @@
     "faker": "^4.1.0",
     "interactjs": "^1.3.4",
     "jest": "^24.8.0",
-    "jest-dom": "^3.5.0",
     "jest-watch-typeahead": "^0.3.1",
     "jsdom": "^11.5.1",
     "moment": "^2.11.1",
@@ -142,8 +144,10 @@
     "prettier-eslint-cli": "^4.7.0",
     "prop-types": "^15.6.2",
     "randomcolor": "^0.5.3",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "react": "^16.10.2",
+    "react-dnd": "^9.3.4",
+    "react-dnd-html5-backend": "^9.3.4",
+    "react-dom": "^16.10.2",
     "react-router-dom": "^4.1.1",
     "react-testing-library": "^6.0.3",
     "regenerator-runtime": "^0.13.2",
