@@ -6,7 +6,7 @@ import {
 } from '../../../__fixtures__/stateAndProps'
 import 'react-testing-library/cleanup-after-each'
 import Timeline from 'lib/Timeline'
-import 'jest-dom/extend-expect'
+import '@testing-library/jest-dom/extend-expect'
 
 /**
  * Testing The Default Functionality
