@@ -1,9 +1,10 @@
 import React from 'react'
 import { render, fireEvent, cleanup } from 'react-testing-library'
-import 'jest-dom/extend-expect'
+import '@testing-library/jest-dom/extend-expect'
 import TimelineMarkers from 'lib/markers/public/TimelineMarkers'
 import CustomMarker from 'lib/markers/public/CustomMarker'
 import { RenderWrapper } from 'test-utility/marker-renderer'
+import { defaultKeys } from '../../../src/lib/default-config'
 
 describe('CustomMarker', () => {
   afterEach(cleanup)
@@ -70,7 +71,8 @@ describe('CustomMarker', () => {
       canvasWidth,
       showPeriod: () => {},
       timelineWidth: 1000,
-      timelineUnit: 'day'
+      timelineUnit: 'day',
+      keys: defaultKeys
     }
 
     const markerDate = now + oneDay / 2
