@@ -1,7 +1,11 @@
 import { calculateScrollCanvas } from 'lib/utility/calendar'
-import {items, groups} from '../../../__fixtures__/itemsAndGroups'
-import {props, state, visibleTimeStart, visibleTimeEnd} from '../../../__fixtures__/stateAndProps'
-
+import { items, groups } from '../../../__fixtures__/itemsAndGroups'
+import {
+  props,
+  state,
+  visibleTimeStart,
+  visibleTimeEnd
+} from '../../../__fixtures__/stateAndProps'
 
 describe('calculateScrollCanvas', () => {
   it('should calculate new scroll state', () => {
@@ -18,7 +22,7 @@ describe('calculateScrollCanvas', () => {
     )
     expect(result).toHaveProperty('visibleTimeStart')
     expect(result).toHaveProperty('visibleTimeEnd')
-    expect(result).toHaveProperty('dimensionItems')
+    expect(result).toHaveProperty('groupsWithItemsDimensions')
   })
   it('should calculate new scroll state correctly', () => {
     const newStartTime = visibleTimeStart + 13 * 60 * 60 * 1000
@@ -35,8 +39,8 @@ describe('calculateScrollCanvas', () => {
     expect(result).toMatchSnapshot()
   })
   it('should skip new calculation if new visible start and visible end in canvas', () => {
-    const newStartTime = visibleTimeStart +  1 * 60 * 60 * 1000
-    const newEndTime = visibleTimeEnd +  1 * 60 * 60 * 1000
+    const newStartTime = visibleTimeStart + 1 * 60 * 60 * 1000
+    const newEndTime = visibleTimeEnd + 1 * 60 * 60 * 1000
     const result = calculateScrollCanvas(
       newStartTime,
       newEndTime,
@@ -48,11 +52,11 @@ describe('calculateScrollCanvas', () => {
     )
     expect(result).toHaveProperty('visibleTimeStart')
     expect(result).toHaveProperty('visibleTimeEnd')
-    expect(result).not.toHaveProperty('dimensionItems')
+    expect(result).not.toHaveProperty('groupsWithItemsDimensions')
   })
   it('should force new calculation', () => {
-    const newStartTime = visibleTimeStart +   1 * 60 * 60 * 1000
-    const newEndTime = visibleTimeEnd +   1 * 60 * 60 * 1000
+    const newStartTime = visibleTimeStart + 1 * 60 * 60 * 1000
+    const newEndTime = visibleTimeEnd + 1 * 60 * 60 * 1000
     const result = calculateScrollCanvas(
       newStartTime.valueOf(),
       newEndTime.valueOf(),
@@ -64,11 +68,11 @@ describe('calculateScrollCanvas', () => {
     )
     expect(result).toHaveProperty('visibleTimeStart')
     expect(result).toHaveProperty('visibleTimeEnd')
-    expect(result).toHaveProperty('dimensionItems')
+    expect(result).toHaveProperty('groupsWithItemsDimensions')
   })
   it('should calculate new state if zoom changed ', () => {
     const newStartTime = visibleTimeStart
-    const newEndTime = visibleTimeEnd +  1 * 60 * 60 * 1000
+    const newEndTime = visibleTimeEnd + 1 * 60 * 60 * 1000
     const result = calculateScrollCanvas(
       newStartTime,
       newEndTime,
@@ -80,11 +84,11 @@ describe('calculateScrollCanvas', () => {
     )
     expect(result).toHaveProperty('visibleTimeStart')
     expect(result).toHaveProperty('visibleTimeEnd')
-    expect(result).toHaveProperty('dimensionItems')
+    expect(result).toHaveProperty('groupsWithItemsDimensions')
   })
   it('should calculate new state if zoom changed correctly', () => {
     const newStartTime = visibleTimeStart
-    const newEndTime = visibleTimeEnd +  1 * 60 * 60 * 1000
+    const newEndTime = visibleTimeEnd + 1 * 60 * 60 * 1000
     const result = calculateScrollCanvas(
       newStartTime,
       newEndTime,
