@@ -4,8 +4,7 @@ import {orderedGroups, dimensionItems} from '../../../__fixtures__/groupOrderAnd
 describe('groupStack', ()=>{
     it('works as expected', ()=>{
         const groupHeight = 0;
-        const totalHeight = 0;
         const index = 0;
-        expect(groupStack(60, dimensionItems[index], dimensionItems, groupHeight, totalHeight, index)).toMatchSnapshot()
+        expect(groupStack(60, dimensionItems[index], dimensionItems, groupHeight, index)).toMatchSnapshot()
     })
 })
