@@ -26,7 +26,7 @@ describe('stackItems', () => {
         state.dragTime,
         state.resizingEdge,
         state.resizeTime,
-        state.newGroupOrder
+        state.newGroupId
       )
     ).toMatchSnapshot()
   })
@@ -47,7 +47,7 @@ describe('stackItems', () => {
         state.dragTime,
         state.resizingEdge,
         state.resizeTime,
-        state.newGroupOrder
+        state.newGroupId
       )
     ).toMatchSnapshot()
   })
@@ -68,7 +68,7 @@ describe('stackItems', () => {
         stateMoveItem.dragTime,
         stateMoveItem.resizingEdge,
         stateMoveItem.resizeTime,
-        stateMoveItem.newGroupOrder
+        stateMoveItem.newGroupId
       )
     ).toMatchSnapshot()
   })
@@ -89,7 +89,7 @@ describe('stackItems', () => {
         stateResizeItemLeft.dragTime,
         stateResizeItemLeft.resizingEdge,
         stateResizeItemLeft.resizeTime,
-        stateResizeItemLeft.newGroupOrder
+        stateResizeItemLeft.newGroupId
       )
     ).toMatchSnapshot()
   })
@@ -110,7 +110,7 @@ describe('stackItems', () => {
         stateResizeItemRight.dragTime,
         stateResizeItemRight.resizingEdge,
         stateResizeItemRight.resizeTime,
-        stateResizeItemRight.newGroupOrder
+        stateResizeItemRight.newGroupId
       )
     ).toMatchSnapshot()
   })
@@ -131,10 +131,10 @@ describe('stackItems', () => {
         state.dragTime,
         state.resizingEdge,
         state.resizeTime,
-        state.newGroupOrder
+        state.newGroupId
       )
     ).toMatchObject({
-      dimensionItems: [],
+      groupsWithItemsDimensions: {},
       height: 0,
       groupHeights: [],
       groupTops: []
