@@ -70,25 +70,25 @@ export default class App extends Component {
     console.log('Context Menu: ' + itemId, moment(time).format())
   }
 
-  handleItemMove = (itemId, dragTime, newGroupOrder) => {
+  handleItemMove = (itemId, dragTime, newGroupId) => {
     const { items, groups } = this.state
 
-    const group = groups[newGroupOrder]
+    const group = groups.find(i => i.id === newGroupId)
 
     this.setState({
       items: items.map(
         item =>
           item.id === itemId
             ? Object.assign({}, item, {
-              start: dragTime,
-              end: dragTime + (item.end - item.start),
-              group: group.id
-            })
+                start: dragTime,
+                end: dragTime + (item.end - item.start),
+                group: group.id
+              })
             : item
       )
     })
 
-    console.log('Moved', itemId, dragTime, newGroupOrder)
+    console.log('Moved', itemId, dragTime, newGroupId)
   }
 
   handleItemResize = (itemId, time, edge) => {
