@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-import React, { Component } from 'react'
+import React, { Component, useEffect } from 'react'
 import moment from 'moment'
 
 import Timeline, {
@@ -10,8 +10,14 @@ import Timeline, {
   CursorMarker,
   CustomHeader,
   SidebarHeader,
-  DateHeader
+  DateHeader,
+  RowColumns,
+  RowItems,
+  GroupRow,
+  HelpersContext
 } from 'react-calendar-timeline'
+import { useDrag, useDrop } from 'react-dnd'
+import * as d3 from 'd3'
 
 import generateFakeData from '../generate-fake-data'
 
@@ -46,12 +52,122 @@ export default class App extends Component {
       .startOf('day')
       .add(1, 'day')
       .toDate()
-
     this.state = {
       groups,
       items,
       defaultTimeStart,
-      defaultTimeEnd
+      defaultTimeEnd,
+      timelineLinks: [],
+      itemsToDrag: [
+        {
+          title: 'print',
+          id: '0a',
+          slots: [
+            {
+              groupId: '1',
+              startTime: moment()
+                .startOf('day')
+                .add(2, 'h'),
+              endTime: moment()
+                .startOf('day')
+                .add(4, 'h')
+            },
+            {
+              groupId: '2',
+              startTime: moment()
+                .startOf('day')
+                .add(8, 'h'),
+              endTime: moment()
+                .startOf('day')
+                .add(16, 'h')
+            }
+          ]
+        },
+        {
+          title: 'cut',
+          id: '1a',
+          slots: [
+            {
+              groupId: '2',
+              startTime: moment()
+                .startOf('day')
+                .add(9, 'h'),
+              endTime: moment()
+                .startOf('day')
+                .add(18, 'h')
+            },
+            {
+              groupId: '5',
+              startTime: moment()
+                .startOf('day')
+                .add(2, 'h'),
+              endTime: moment()
+                .startOf('day')
+                .add(10, 'h')
+            }
+          ]
+        },
+        {
+          title: 'fold',
+          id: '2a',
+          slots: [
+            {
+              groupId: '4',
+              startTime: moment()
+                .startOf('day')
+                .add(9, 'h'),
+              endTime: moment()
+                .startOf('day')
+                .add(18, 'h')
+            },
+            {
+              groupId: '6',
+              startTime: moment()
+                .startOf('day')
+                .add(2, 'h'),
+              endTime: moment()
+                .startOf('day')
+                .add(8, 'h')
+            }
+          ]
+        }
+      ],
+      unavailableSlots: {
+        '1': [
+          {
+            id: '0',
+            groupId: '0',
+            startTime: moment()
+              .startOf('day')
+              .add(16, 'h'),
+            endTime: moment()
+              .startOf('day')
+              .add(20, 'h')
+          }
+        ],
+        '3': [
+          {
+            id: '1',
+            groupId: '3',
+            startTime: moment()
+              .startOf('day')
+              .add(13, 'h'),
+            endTime: moment()
+              .startOf('day')
+              .add(15, 'h')
+          },
+          {
+            id: '2',
+            groupId: '3',
+            startTime: moment()
+              .startOf('day')
+              .add(22, 'h'),
+            endTime: moment()
+              .startOf('day')
+              .add(24, 'h')
+          }
+        ]
+      }
     }
   }
 
@@ -67,11 +183,29 @@ export default class App extends Component {
     console.log('Canvas context menu', group, moment(time).format())
   }
 
+  findItemById = itemId => {
+    return this.state.items.find(i => i.id === itemId)
+  }
+
+  tempItemId = undefined
+
   handleItemClick = (itemId, _, time) => {
     console.log('Clicked: ' + itemId, moment(time).format())
   }
 
   handleItemSelect = (itemId, _, time) => {
+    if (!this.tempItemId) {
+      this.tempItemId = itemId
+    } else {
+      this.setState(
+        state => ({
+          timelineLinks: [...state.timelineLinks, [this.tempItemId, itemId]]
+        }),
+        () => {
+          this.tempItemId = undefined
+        }
+      )
+    }
     console.log('Selected: ' + itemId, moment(time).format())
   }
 
@@ -83,10 +217,10 @@ export default class App extends Component {
     console.log('Context Menu: ' + itemId, moment(time).format())
   }
 
-  handleItemMove = (itemId, dragTime, newGroupOrder) => {
+  handleItemMove = (itemId, dragTime, newGroupId) => {
     const { items, groups } = this.state
 
-    const group = groups[newGroupOrder]
+    const group = groups.find(i => i.id === newGroupId)
 
     this.setState({
       items: items.map(
@@ -101,7 +235,7 @@ export default class App extends Component {
       )
     })
 
-    console.log('Moved', itemId, dragTime, newGroupOrder)
+    console.log('Moved', itemId, dragTime, newGroupId)
   }
 
   handleItemResize = (itemId, time, edge) => {
@@ -136,13 +270,83 @@ export default class App extends Component {
   }
 
   moveResizeValidator = (action, item, time) => {
-    if (time < new Date().getTime()) {
-      var newTime =
-        Math.ceil(new Date().getTime() / (15 * 60 * 1000)) * (15 * 60 * 1000)
-      return newTime
+    const unavailableSlots = this.state.unavailableSlots[item.group]
+    const originalItem = this.state.items.find(i => i.id === item.id)
+    const startTimeInMoment = moment(time, 'x')
+    const endTimeInMoment = moment(item.end - item.start + time, 'x')
+    if (unavailableSlots) {
+      const violation = unavailableSlots.find(slot => {
+        const { startTime, endTime } = slot
+        console.log(endTimeInMoment.format(), startTime.format())
+        return (
+          (startTimeInMoment.isAfter(startTime) &&
+            startTimeInMoment.isBefore(endTime)) ||
+          (endTimeInMoment.isAfter(startTime) &&
+            endTimeInMoment.isBefore(endTime))
+        )
+      })
+      if (violation)
+        return (
+          violation.startTime.valueOf() -
+          (originalItem.end - originalItem.start)
+        )
     }
-
     return time
+  }
+
+  handleDrop = (item, slot) => {
+    const fullItem = this.state.itemsToDrag.find(i => i.id === item.id)
+    this.setState(state => ({
+      itemsToDrag: state.itemsToDrag.filter(i => item.id !== i.id),
+      items: [
+        ...state.items,
+        {
+          title: fullItem.title,
+          id: item.id,
+          group: slot.groupId,
+          start: slot.startTime.valueOf(),
+          end: slot.endTime.valueOf()
+        }
+      ]
+    }))
+  }
+
+  rowRenderer = ({
+    rowData,
+    getLayerRootProps,
+    group,
+    itemsWithInteractions
+  }) => {
+    const helpers = React.useContext(HelpersContext)
+    const { itemsToDrag, unavailableSlots, timelineLinks } = rowData
+    const groupUnavailableSlots = unavailableSlots[group.id]
+      ? unavailableSlots[group.id]
+      : []
+    return (
+      <GroupRow>
+        <RowItems />
+        <UnavailableLayer
+          getLayerRootProps={getLayerRootProps}
+          getLeftOffsetFromDate={helpers.getLeftOffsetFromDate}
+          groupUnavailableSlots={groupUnavailableSlots}
+        />
+        <DroppablesLayer
+          getLayerRootProps={getLayerRootProps}
+          itemsToDrag={itemsToDrag}
+          getLeftOffsetFromDate={helpers.getLeftOffsetFromDate}
+          handleDrop={this.handleDrop}
+          group={group}
+        />
+        <Links
+          timelineLinks={timelineLinks}
+          getItemAbsoluteDimensions={helpers.getItemAbsoluteDimensions}
+          getItemDimensions={helpers.getItemDimensions}
+          group={group}
+          getLayerRootProps={getLayerRootProps}
+          items={itemsWithInteractions}
+        />
+      </GroupRow>
+    )
   }
 
   render() {
@@ -156,7 +360,7 @@ export default class App extends Component {
         sidebarWidth={150}
         sidebarContent={<div>Above The Left</div>}
         canMove
-        canResize="right"
+        canResize="both"
         canSelect
         itemsSorted
         itemTouchSendsClick={false}
@@ -174,9 +378,68 @@ export default class App extends Component {
         onItemResize={this.handleItemResize}
         onItemDoubleClick={this.handleItemDoubleClick}
         onTimeChange={this.handleTimeChange}
-        moveResizeValidator={this.moveResizeValidator}
+        rowRenderer={this.rowRenderer}
+        rowData={{
+          itemsToDrag: this.state.itemsToDrag,
+          unavailableSlots: this.state.unavailableSlots,
+          timelineLinks: this.state.timelineLinks
+        }}
+        // moveResizeValidator={this.moveResizeValidator}
       >
-        <TimelineMarkers>
+        <TimelineHeaders>
+          <DateHeader />
+          <CustomHeader
+            headerData={{
+              itemsToDrag: this.state.itemsToDrag
+            }}
+          >
+            {({
+              headerContext: { intervals },
+              getRootProps,
+              getIntervalProps,
+              showPeriod,
+              data: { itemsToDrag }
+            }) => {
+              const {getLeftOffsetFromDate} = React.useContext(HelpersContext)
+              return (
+                <div {...getRootProps()}>
+                  <div
+                    style={{
+                      height: '100%',
+                      transform: `translateX(${getLeftOffsetFromDate(
+                        moment()
+                          .startOf('day')
+                          .valueOf()
+                      )}px)`,
+                      display: 'flex'
+                    }}
+                  >
+                    {itemsToDrag.map(dragItem => {
+                      return (
+                        <Draggable
+                          key={dragItem.id}
+                          id={dragItem.id}
+                          style={{
+                            height: '100%',
+                            width: 100,
+                            background: 'white',
+                            marginLeft: 15,
+                            border: '1px solid white'
+                          }}
+                          onDragEnd={item => console.log('dragEnd', item)}
+                          onDragStart={item => console.log('dragStart', item)}
+                        >
+                          {dragItem.title}
+                        </Draggable>
+                      )
+                    })}
+                  </div>
+                </div>
+              )
+            }}
+          </CustomHeader>
+        </TimelineHeaders>
+        {/* <TimelineMarkers>
           <TodayMarker />
           <CustomMarker
             date={
@@ -197,8 +460,205 @@ export default class App extends Component {
             }}
           </CustomMarker>
           <CursorMarker />
-        </TimelineMarkers>
+        </TimelineMarkers> */}
       </Timeline>
     )
   }
 }
+
+function Link({
+  timelineLink,
+  getItemAbsoluteDimensions,
+  getItemDimensions,
+  group,
+  items
+}) {
+  const [startId, endId] = timelineLink
+  const startItem = items.find(i => i.id === startId)
+  if (startItem.group !== group.id) return null
+  const startItemDimensions = getItemAbsoluteDimensions(startId) || {
+    left: 0,
+    top: 0
+  }
+  const endItemDimensions = getItemAbsoluteDimensions(endId) || {
+    left: 0,
+    top: 0
+  }
+  let startLink = [startItemDimensions.left, startItemDimensions.top]
+  let endLink = [endItemDimensions.left, endItemDimensions.top]
+  const isEndLinkBeforeStart =
+    endLink[0] <= startLink[0] || endLink[1] <= startLink[1]
+  let itemLink = isEndLinkBeforeStart
+    ? [endLink, startLink]
+    : [startLink, endLink]
+  const lineGenerator = d3.line()
+  const [startLk, endLk] = itemLink
+  const endPoint = [endLk[0] - startLk[0], endLk[1] - startLk[1]]
+  const itemDimensions = getItemDimensions(startId)
+  return (
+    <svg
+      style={{
+        position: 'absolute',
+        left: startLk[0],
+        zIndex: 200,
+        top: itemDimensions.top,
+        height: endPoint[1],
+        //handle case where endPoint is 0
+        width: endPoint[0] > 2 ? endPoint[0] : 2,
+        pointerEvents: 'none'
+      }}
+    >
+      <path
+        d={lineGenerator([[0, 0], endPoint])}
+        stroke="red"
+        strokeWidth="2"
+        fill="none"
+      />
+    </svg>
+  )
+}
+
+const Links = React.memo(({
+  timelineLinks,
+  getItemAbsoluteDimensions,
+  getItemDimensions,
+  group,
+  getLayerRootProps,
+  items
+}) => {
+  return (
+    <div {...getLayerRootProps()}>
+      {timelineLinks.map((timelineLink, i) => {
+        const [startId, endId] = timelineLink
+        return (
+          <Link
+            timelineLink={timelineLink}
+            key={`${startId}${endId}`}
+            getItemAbsoluteDimensions={getItemAbsoluteDimensions}
+            getItemDimensions={getItemDimensions}
+            group={group}
+            items={items}
+          />
+        )
+      })}
+    </div>
+  )
+})
+
+function Droppable({ children, itemIdAccepts, style, slot, onDrop, ...rest }) {
+  const [collected, droppableRef] = useDrop({
+    drop: (item, monitor) => {
+      onDrop(item, slot)
+    },
+    accept: itemIdAccepts,
+    collect: monitor => ({
+      canDrop: monitor.canDrop()
+    })
+  })
+  const isVisable = collected.canDrop
+  return (
+    <div
+      style={{
+        ...style,
+        display: isVisable ? 'initial' : 'none'
+      }}
+      ref={droppableRef}
+      {...rest}
+    >
+      {children}
+    </div>
+  )
+}
+
+function Draggable({ id, children, onDragStart, onDragEnd, ...rest }) {
+  const [collectedProps, dragRef] = useDrag({
+    item: { id, type: id },
+    begin: monitor => {
+      onDragStart(id)
+    },
+    end: (item, monitor) => {
+      console.log(monitor)
+      onDragEnd(item)
+    }
+  })
+  return (
+    <div {...rest} ref={dragRef}>
+      {children}
+    </div>
+  )
+}
+
+function DroppablesLayer({
+  getLayerRootProps,
+  itemsToDrag,
+  getLeftOffsetFromDate,
+  handleDrop,
+  group
+}) {
+  return (
+    <div {...getLayerRootProps()}>
+      {itemsToDrag.map((item, index) => {
+        return item.slots
+          .filter(slot => slot.groupId === group.id)
+          .map(slot => {
+            const left = getLeftOffsetFromDate(slot.startTime.valueOf())
+            const right = getLeftOffsetFromDate(slot.endTime.valueOf())
+            return (
+              <Droppable
+                key={index}
+                style={{
+                  position: 'absolute',
+                  left: left,
+                  width: right - left,
+                  backgroundColor: 'purple',
+                  height: '100%'
+                }}
+                itemIdAccepts={item.id}
+                slot={slot}
+                onDrop={handleDrop}
+              >
+                {item.title}
+              </Droppable>
+            )
+          })
+      })}
+    </div>
+  )
+}
+
+const UnavailableLayer = (({
+  getLayerRootProps,
+  groupUnavailableSlots,
+  getLeftOffsetFromDate
+}) => {
+  useEffect(() => {
+    return () => {
+      // console.log("unmount UnavailableLayer")
+    }
+  })
+  return (
+    <div {...getLayerRootProps()}>
+      {groupUnavailableSlots.map(slot => {
+        const left = getLeftOffsetFromDate(slot.startTime.valueOf())
+        const right = getLeftOffsetFromDate(slot.endTime.valueOf())
+        return (
+          <div
+            key={slot.id}
+            style={{
+              position: 'absolute',
+              left: left,
+              width: right - left,
+              backgroundColor: '#ECF0F1',
+              height: '100%',
+              display: 'flex',
+              alignItems: 'center',
+              justifyContent: 'center'
+            }}
+          >
+            <span style={{ height: 12 }}>unavailable</span>
+          </div>
+        )
+      })}
+    </div>
+  )
+})
