@@ -1,9 +1,9 @@
 import faker from 'faker'
 import randomColor from 'randomcolor'
 import moment from 'moment'
-
+faker.seed(30);
 export default function (groupCount = 30, itemCount = 1000, daysInPast = 30) {
-  let randomSeed = Math.floor(Math.random() * 1000)
+  let randomSeed = 10
   let groups = []
   for (let i = 0; i < groupCount; i++) {
     groups.push({
@@ -27,8 +27,8 @@ export default function (groupCount = 30, itemCount = 1000, daysInPast = 30) {
       title: faker.hacker.phrase(),
       start: startValue,
       end: endValue,
-      canMove: startValue > new Date().getTime(),
-      canResize: startValue > new Date().getTime() ? (endValue > new Date().getTime() ? 'both' : 'left') : (endValue > new Date().getTime() ? 'right' : false),
+      // canMove: startValue > new Date().getTime(),
+      // canResize: startValue > new Date().getTime() ? (endValue > new Date().getTime() ? 'both' : 'left') : (endValue > new Date().getTime() ? 'right' : false),
       className: (moment(startDate).day() === 6 || moment(startDate).day() === 0) ? 'item-weekend' : '',
       bgColor: randomColor({ luminosity: 'light', seed: randomSeed + i, format:'rgba', alpha:0.6 }),
       selectedBgColor: randomColor({ luminosity: 'light', seed: randomSeed + i, format:'rgba', alpha:1 }),
