@@ -1,10 +1,9 @@
 import PropTypes from 'prop-types'
-import React, { Component } from 'react'
+import React, { PureComponent, Component } from 'react'
 import Item from './Item'
-// import ItemGroup from './ItemGroup'
-
-import { _get, arraysEqual, keyBy } from '../utility/generic'
-import { getGroupOrders, getVisibleItems } from '../utility/calendar'
+import { TimelineStateConsumer } from '../timeline/TimelineStateContext'
+import { ItemsConsumer } from './ItemsContext'
+import { _get, arraysEqual } from '../utility/generic'
 
 const canResizeLeft = (item, canResize) => {
   const value =
@@ -18,14 +17,16 @@ const canResizeRight = (item, canResize) => {
   return value === 'right' || value === 'both' || value === true
 }
 
-export default class Items extends Component {
+export class Items extends Component {
   static propTypes = {
-    groups: PropTypes.oneOfType([PropTypes.array, PropTypes.object]).isRequired,
     items: PropTypes.oneOfType([PropTypes.array, PropTypes.object]).isRequired,
 
     canvasTimeStart: PropTypes.number.isRequired,
     canvasTimeEnd: PropTypes.number.isRequired,
     canvasWidth: PropTypes.number.isRequired,
+    visibleTimeEnd: PropTypes.number.isRequired,
+    visibleTimeStart: PropTypes.number.isRequired,
+    timelineWidth: PropTypes.number.isRequired,
 
     dragSnap: PropTypes.number,
     minResizeWidth: PropTypes.number,
@@ -51,10 +52,18 @@ export default class Items extends Component {
     itemRenderer: PropTypes.func,
     selected: PropTypes.array,
 
-    dimensionItems: PropTypes.array,
-    groupTops: PropTypes.array,
+    groupDimensions: PropTypes.object,
     useResizeHandle: PropTypes.bool,
-    scrollRef: PropTypes.object
+    scrollRef: PropTypes.object,
+    order: PropTypes.object,
+
+    onDragStart: PropTypes.func.isRequired,
+    onDragEnd: PropTypes.func.isRequired,
+    onResizeStart:  PropTypes.func.isRequired,
+    dragging: PropTypes.bool.isRequired,
+    dragOffset: PropTypes.number.isRequired,
+    interactingItemId: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+    resizeEdge: PropTypes.oneOf(['right', 'left']),
   }
 
   static defaultProps = {
@@ -63,9 +72,8 @@ export default class Items extends Component {
 
   shouldComponentUpdate(nextProps) {
     return !(
-      arraysEqual(nextProps.groups, this.props.groups) &&
       arraysEqual(nextProps.items, this.props.items) &&
-      arraysEqual(nextProps.dimensionItems, this.props.dimensionItems) &&
+      nextProps.groupDimensions === this.props.groupDimensions &&
       nextProps.keys === this.props.keys &&
       nextProps.canvasTimeStart === this.props.canvasTimeStart &&
       nextProps.canvasTimeEnd === this.props.canvasTimeEnd &&
@@ -77,7 +85,11 @@ export default class Items extends Component {
       nextProps.canChangeGroup === this.props.canChangeGroup &&
       nextProps.canMove === this.props.canMove &&
       nextProps.canResize === this.props.canResize &&
-      nextProps.canSelect === this.props.canSelect     
+      nextProps.canSelect === this.props.canSelect &&
+      nextProps.dragging === this.props.dragging &&
+      nextProps.resizing === this.props.resizing &&
+      nextProps.resizeEdge === this.props.resizeEdge &&
+      nextProps.interactingItemId === this.props.interactingItemId
     )
   }
 
@@ -90,81 +102,114 @@ export default class Items extends Component {
     }
   }
 
-  getVisibleItems(canvasTimeStart, canvasTimeEnd) {
-    const { keys, items } = this.props
-
-    return getVisibleItems(items, canvasTimeStart, canvasTimeEnd, keys)
+  isInteractingItem = (item) => {
+    return this.props.interactingItemId === _get(item, this.props.keys.itemIdKey)
   }
 
   render() {
     const {
-      canvasTimeStart,
-      canvasTimeEnd,
-      dimensionItems,
       keys,
-      groups
+      groupDimensions,
+      order,
+      items
     } = this.props
-    const { itemIdKey, itemGroupKey } = keys
-
-    const groupOrders = getGroupOrders(groups, keys)
-    const visibleItems = this.getVisibleItems(
-      canvasTimeStart,
-      canvasTimeEnd,
-      groupOrders
-    )
-    const sortedDimensionItems = keyBy(dimensionItems, 'id')
-
+    const { itemIdKey } = keys
     return (
       <div className="rct-items">
-        {visibleItems
-          .filter(item => sortedDimensionItems[_get(item, itemIdKey)])
-          .map(item => (
-            <Item
-              key={_get(item, itemIdKey)}
-              item={item}
-              keys={this.props.keys}
-              order={groupOrders[_get(item, itemGroupKey)]}
-              dimensions={
-                sortedDimensionItems[_get(item, itemIdKey)].dimensions
-              }
-              selected={this.isSelected(item, itemIdKey)}
-              canChangeGroup={
-                _get(item, 'canChangeGroup') !== undefined
-                  ? _get(item, 'canChangeGroup')
-                  : this.props.canChangeGroup
-              }
-              canMove={
-                _get(item, 'canMove') !== undefined
-                  ? _get(item, 'canMove')
-                  : this.props.canMove
-              }
-              canResizeLeft={canResizeLeft(item, this.props.canResize)}
-              canResizeRight={canResizeRight(item, this.props.canResize)}
-              canSelect={
-                _get(item, 'canSelect') !== undefined
-                  ? _get(item, 'canSelect')
-                  : this.props.canSelect
-              }
-              useResizeHandle={this.props.useResizeHandle}
-              groupTops={this.props.groupTops}
-              canvasTimeStart={this.props.canvasTimeStart}
-              canvasTimeEnd={this.props.canvasTimeEnd}
-              canvasWidth={this.props.canvasWidth}
-              dragSnap={this.props.dragSnap}
-              minResizeWidth={this.props.minResizeWidth}
-              onResizing={this.props.itemResizing}
-              onResized={this.props.itemResized}
-              moveResizeValidator={this.props.moveResizeValidator}
-              onDrag={this.props.itemDrag}
-              onDrop={this.props.itemDrop}
-              onItemDoubleClick={this.props.onItemDoubleClick}
-              onContextMenu={this.props.onItemContextMenu}
-              onSelect={this.props.itemSelect}
-              itemRenderer={this.props.itemRenderer}
-              scrollRef={this.props.scrollRef}
-            />
-          ))}
+        {items.map((item, i) => {
+          const isInteractingItem = this.isInteractingItem(item)
+          return (
+          <Item
+            key={_get(item, itemIdKey)}
+            item={item}
+            keys={this.props.keys}
+            order={order}
+            dimensions={groupDimensions.itemDimensions[i].dimensions}
+            selected={this.isSelected(item, itemIdKey)}
+            canChangeGroup={
+              _get(item, 'canChangeGroup') !== undefined
+                ? _get(item, 'canChangeGroup')
+                : this.props.canChangeGroup
+            }
+            canMove={
+              _get(item, 'canMove') !== undefined
+                ? _get(item, 'canMove')
+                : this.props.canMove
+            }
+            canResizeLeft={canResizeLeft(item, this.props.canResize)}
+            canResizeRight={canResizeRight(item, this.props.canResize)}
+            canSelect={
+              _get(item, 'canSelect') !== undefined
+                ? _get(item, 'canSelect')
+                : this.props.canSelect
+            }
+            useResizeHandle={this.props.useResizeHandle}
+            canvasTimeStart={this.props.canvasTimeStart}
+            canvasTimeEnd={this.props.canvasTimeEnd}
+            canvasWidth={this.props.canvasWidth}
+            dragSnap={this.props.dragSnap}
+            minResizeWidth={this.props.minResizeWidth}
+            onResizing={this.props.itemResizing}
+            onResized={this.props.itemResized}
+            moveResizeValidator={this.props.moveResizeValidator}
+            onDrag={this.props.itemDrag}
+            onDrop={this.props.itemDrop}
+            onItemDoubleClick={this.props.onItemDoubleClick}
+            onContextMenu={this.props.onItemContextMenu}
+            onSelect={this.props.itemSelect}
+            itemRenderer={this.props.itemRenderer}
+            scrollRef={this.props.scrollRef}
+            dragging={isInteractingItem && this.props.dragging}
+            resizing={isInteractingItem && this.props.resizing}
+            dragOffset={isInteractingItem ? this.props.dragOffset: 0}
+            resizeEdge={isInteractingItem ? this.props.resizeEdge: undefined}
+            onDragStart={this.props.onDragStart}
+            onDragEnd={this.props.onDragEnd}
+            onResizeStart={this.props.onResizeStart}
+            visibleTimeEnd={this.props.visibleTimeEnd}
+            visibleTimeStart={this.props.visibleTimeStart}
+            timelineWidth={this.props.timelineWidth}
+          />
+        )})}
       </div>
     )
   }
 }
+
+class ItemsWrapper extends PureComponent {
+  render() {
+    return (
+      <TimelineStateConsumer>
+        {({ getTimelineState }) => {
+          const {
+            canvasTimeStart,
+            canvasTimeEnd,
+            canvasWidth,
+            visibleTimeEnd,
+            visibleTimeStart,
+            timelineWidth,
+            keys
+          } = getTimelineState()
+          return (
+            <ItemsConsumer>
+              {props => (
+                <Items
+                  canvasTimeStart={canvasTimeStart}
+                  canvasTimeEnd={canvasTimeEnd}
+                  canvasWidth={canvasWidth}
+                  visibleTimeEnd={visibleTimeEnd}
+                  visibleTimeStart={visibleTimeStart}
+                  timelineWidth={timelineWidth}
+                  keys={keys}
+                  {...props}
+                />
+              )}
+            </ItemsConsumer>
+          )
+        }}
+      </TimelineStateConsumer>
+    )
+  }
+}
+
+export default ItemsWrapper
