@@ -30,7 +30,9 @@ const defaultContextState = {
 }
 /* eslint-enable */
 
-const { Consumer, Provider } = createReactContext(defaultContextState)
+const TimelineStateContext = createReactContext(defaultContextState)
+
+const { Consumer, Provider } = TimelineStateContext
 
 export class TimelineStateProvider extends React.Component {
   /* eslint-disable react/no-unused-prop-types */
@@ -44,6 +46,7 @@ export class TimelineStateProvider extends React.Component {
     showPeriod: PropTypes.func.isRequired,
     timelineUnit: PropTypes.string.isRequired,
     timelineWidth: PropTypes.number.isRequired,
+    keys: PropTypes.object.isRequired
   }
 
   constructor(props) {
@@ -68,6 +71,7 @@ export class TimelineStateProvider extends React.Component {
       canvasWidth,
       timelineUnit,
       timelineWidth,
+      keys,
     } = this.props
     return {
       visibleTimeStart,
@@ -77,6 +81,7 @@ export class TimelineStateProvider extends React.Component {
       canvasWidth,
       timelineUnit,
       timelineWidth,
+      keys,
     } // REVIEW,
   }
 
@@ -110,3 +115,4 @@ export class TimelineStateProvider extends React.Component {
 }
 
 export const TimelineStateConsumer = Consumer
+export default TimelineStateContext;
